Add a test
Because a promotion starts with a price reduction and lasts thirty days,
and because a price reduction only starts a promotion if the price has
been stable for thirty days, it follows that further price reductions
during a promotion will neither extend that promotion nor override it
with a new one. Add a test clarifying this.